refactor: move package checker to the top level
